[plugin-mobile-app] Make JavaScript executing steps available for mobile app tests
