+++ vividus-extension-selenium/src/main/java/org/vividus/steps/ui/AbstractExecuteScriptSteps.java
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2019-2022 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.vividus.steps.ui;
+
+import java.util.Set;
+import java.util.function.Supplier;
+
+import org.vividus.context.VariableContext;
+import org.vividus.softassert.ISoftAssert;
+import org.vividus.variable.VariableScope;
+
+@SuppressWarnings("PMD.AbstractClassWithoutAbstractMethod")
+public abstract class AbstractExecuteScriptSteps
+{
+    private final ISoftAssert softAssert;
+    private final VariableContext variableContext;
+
+    protected AbstractExecuteScriptSteps(ISoftAssert softAssert, VariableContext variableContext)
+    {
+        this.softAssert = softAssert;
+        this.variableContext = variableContext;
+    }
+
+    protected void assertAndSaveResult(Supplier<Object> resultProvider, Set<VariableScope> scopes, String variableName)
+    {
+        Object result = resultProvider.get();
+        if (softAssert.assertNotNull("Returned result is not null", result))
+        {
+            variableContext.putVariable(scopes, variableName, result);
+        }
+    }
+}
