+++ vividus-extension-selenium/src/main/java/org/vividus/steps/ui/web/model/JsArgumentType.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019-2021 the original author or authors.
+ * Copyright 2019-2022 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,31 +16,30 @@
 
 package org.vividus.steps.ui.web.model;
 
-import org.jbehave.core.annotations.AsParameters;
+import java.util.Map;
 
-@AsParameters
-public class JsArgument
+import org.vividus.util.json.JsonUtils;
+
+public enum JsArgumentType
 {
-    private String value;
-    private JsArgumentType type;
-
-    public String getValue()
+    STRING
     {
-        return value;
-    }
-
-    public void setValue(String value)
+        @Override
+        public Object convert(String object)
+        {
+            return object;
+        }
+    },
+    OBJECT
     {
-        this.value = value;
-    }
+        private final transient JsonUtils jsonUtils = new JsonUtils();
 
-    public JsArgumentType getType()
-    {
-        return type;
-    }
+        @Override
+        public Object convert(String object)
+        {
+            return jsonUtils.toObject(object, Map.class);
+        }
+    };
 
-    public void setType(JsArgumentType type)
-    {
-        this.type = type;
-    }
+    public abstract Object convert(String object);
 }
